@@ -1,16 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace localization.Models.Example
 {
-    public class ParameterViewModel
-    {
-        public int Index { get; set; }
+	public class ParameterViewModel
+	{
+		public int Index { get; set; }
 
-        [Display(Name = "Test parameter!")]
-        public string Test { get; set; }
-    }
+		[Display(Name = "Test parameter!")]
+		public string Test { get; set; }
+	}
 }
