@@ -1,6 +1,6 @@
-﻿@{
-    ViewData["Title"] = Localizer["Example Index"];
+@{
+    ViewData["Title"] = "Example Index";
 }
 
-@Localizer["Hello world!"]
-<a asp-controller="home" asp-action="index" cms-culture="@ViewData["culture"]">@Localizer["English"]</a>
+hello world
+<a asp-controller="home" asp-action="index">test</a>
