@@ -1,9 +1,3 @@
-@using Microsoft.AspNetCore.Identity
-@using localization
+﻿@using localization
 @using localization.Models
-@using localization.Models.AccountViewModels
-@using localization.Models.ManageViewModels
-@using Microsoft.AspNetCore.Mvc.Localization
-@inject IViewLocalizer Localizer
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper "*, localization"
